+++ src/components/crystal-toolkit/CrystalToolkitAnimationScene/index.tsx
@@ -1,1 +1,2 @@
 export { CrystalToolkitAnimationScene } from './CrystalToolkitAnimationScene';
+export type { CrystalToolkitAnimationSceneProps } from './CrystalToolkitAnimationScene';