+++ src/components/crystal-toolkit/CrystalToolkitScene/index.tsx
@@ -1,1 +1,2 @@
 export { CrystalToolkitScene } from './CrystalToolkitScene';
+export type { CrystalToolkitSceneProps } from './CrystalToolkitScene';