+++ src/components/crystal-toolkit/PhononAnimationScene/index.tsx
@@ -1,1 +1,2 @@
 export { PhononAnimationScene } from './PhononAnimationScene';
+export type { PhononAnimationSceneProps } from './PhononAnimationScene';