+++ src/index.ts
@@ -4,3 +4,4 @@ export * from './components/crystal-toolkit/CrystalToolkitScene';
 export * from './components/crystal-toolkit/Download';
 export * from './components/crystal-toolkit/PhononAnimationScene';
 export { default as Scene } from './components/crystal-toolkit/scene/Scene';
+export type { PrimitiveMode, SelectionMode } from './components/crystal-toolkit/scene/constants';